datapath: remove 'unparallel' build tag in tests
We've gone through great lengths to avoid using build tags in tests since it
messes with editor tooling, code completion, go-to reference, etc.

Use the testutils.SerializedTest helper introduced in a prior commit. Remove
the 'unparallel' build tag.
