+++ pkg/datapath/linux/ipsec/xfrm_state_cache_test.go
@@ -1,5 +1,3 @@
-//go:build unparallel
-
 // SPDX-License-Identifier: Apache-2.0
 // Copyright Authors of Cilium
